Share common CodeTF models where possible
