+++ src/codemodder/codetf/common.py
@@ -1,8 +1,9 @@
 from abc import ABCMeta
 from enum import Enum
 from pathlib import Path
+from typing import Optional
 
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict, model_validator
 
 from codemodder.logging import logger
 
@@ -26,3 +27,60 @@ class CodeTFWriter(BaseModel, metaclass=ABCMeta):
             return 2
         logger.debug("wrote report to %s", outfile)
         return 0
+
+
+class Rule(BaseModel):
+    id: str
+    name: str
+    url: Optional[str] = None
+
+    model_config = ConfigDict(frozen=True)
+
+
+class Finding(BaseModel):
+    id: str
+    rule: Rule
+
+    model_config = ConfigDict(frozen=True)
+
+
+class Action(CaseInsensitiveEnum):
+    ADD = "add"
+    REMOVE = "remove"
+
+
+class PackageResult(CaseInsensitiveEnum):
+    COMPLETED = "completed"
+    FAILED = "failed"
+    SKIPPED = "skipped"
+
+
+class DiffSide(CaseInsensitiveEnum):
+    LEFT = "left"
+    RIGHT = "right"
+
+
+class PackageAction(BaseModel):
+    action: Action
+    result: PackageResult
+    package: str
+
+
+class Change(BaseModel):
+    lineNumber: int
+    description: Optional[str]
+    diffSide: DiffSide = DiffSide.RIGHT
+    properties: Optional[dict] = None
+    packageActions: Optional[list[PackageAction]] = None
+
+    @model_validator(mode="after")
+    def validate_lineNumber(self):
+        if self.lineNumber < 1:
+            raise ValueError("lineNumber must be greater than 0")
+        return self
+
+    @model_validator(mode="after")
+    def validate_description(self):
+        if self.description is not None and not self.description:
+            raise ValueError("description must not be empty")
+        return self
