+++ src/codemodder/codetf/v3/codetf.py
@@ -5,7 +5,8 @@ from typing import Optional
 
 from pydantic import BaseModel, model_validator
 
-from ..common import CaseInsensitiveEnum, CodeTFWriter
+from ..common import Change, CodeTFWriter, Finding
+from ..v2.codetf import Finding as V2Finding
 
 
 class Run(BaseModel):
@@ -42,59 +43,6 @@ class FixStatus(BaseModel):
     status: FixStatusType
     reason: Optional[str]
     details: Optional[str]
-
-
-class Rule(BaseModel):
-    id: str
-    name: str
-    url: Optional[str] = None
-
-
-class Finding(BaseModel):
-    id: str
-    rule: Optional[Rule] = None
-
-
-class Action(CaseInsensitiveEnum):
-    ADD = "add"
-    REMOVE = "remove"
-
-
-class PackageResult(CaseInsensitiveEnum):
-    COMPLETED = "completed"
-    FAILED = "failed"
-    SKIPPED = "skipped"
-
-
-class DiffSide(CaseInsensitiveEnum):
-    LEFT = "left"
-    RIGHT = "right"
-
-
-class PackageAction(BaseModel):
-    action: Action
-    result: PackageResult
-    package: str
-
-
-class Change(BaseModel):
-    lineNumber: int
-    description: Optional[str]
-    diffSide: DiffSide = DiffSide.RIGHT
-    properties: Optional[dict] = None
-    packageActions: Optional[list[PackageAction]] = None
-
-    @model_validator(mode="after")
-    def validate_lineNumber(self):
-        if self.lineNumber < 1:
-            raise ValueError("lineNumber must be greater than 0")
-        return self
-
-    @model_validator(mode="after")
-    def validate_description(self):
-        if self.description is not None and not self.description:
-            raise ValueError("description must not be empty")
-        return self
 
 
 class ChangeSet(BaseModel):
@@ -158,7 +106,7 @@ class FixQuality(BaseModel):
 class FixResult(BaseModel):
     """Result corresponding to a single finding"""
 
-    finding: Finding
+    finding: Finding | V2Finding
     fixStatus: FixStatus
     changeSets: list[ChangeSet]
     fixMetadata: Optional[FixMetadata] = None
